Converted double to float.

+++ cublas_mm/cusparse_mm_kernel.h
@@ -15,9 +15,9 @@
 // (m x k) * (k * n) = (m x n)
 // note: row_ind.len = lda + 1
 
-void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
+void cusparse_mm_wrapper(float *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
                          int nnzA, int h_A_rowptr_size,
-                         double *h_B_dense, int h_B_rows, int h_B_cols)
+                         float *h_B_dense, int h_B_rows, int h_B_cols)
 {
     // Initialize cuSPARSE
     cusparseHandle_t handle;
@@ -27,11 +27,11 @@ void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
     const int n = h_B_cols;
 
     // Host side dense matrices
-    double *h_C_dense = (double *)malloc(m * n * sizeof(*h_C_dense));
+    float *h_C_dense = (float *)malloc(m * n * sizeof(*h_C_dense));
     // Create device arrays and copy host arrays to them
-    double *d_B_dense;
+    float *d_B_dense;
     gpuErrchk(cudaMalloc(&d_B_dense, k * n * sizeof(*d_B_dense)));
-    double *d_C_dense;
+    float *d_C_dense;
     gpuErrchk(cudaMalloc(&d_C_dense, m * n * sizeof(*d_C_dense)));
 
     // copy B from host to device
@@ -63,7 +63,7 @@ void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
     gpuErrchk(cudaMemcpy(h_nnzPerVectorB, d_nnzPerVectorB, k * sizeof(*h_nnzPerVectorB), cudaMemcpyDeviceToHost));
 
     // Device side sparse matrix A
-    double *d_A;
+    float *d_A;
     gpuErrchk(cudaMalloc(&d_A, nnzA * sizeof(*d_A)));
     int *d_A_RowIndices;
     gpuErrchk(cudaMalloc(&d_A_RowIndices, (m + 1) * sizeof(*d_A_RowIndices)));
@@ -75,7 +75,7 @@ void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
     cudaMemcpy(d_A_ColIndices, h_A_ColIndices, nnzA * sizeof(*d_A_ColIndices), cudaMemcpyHostToDevice);
 
     // Device side sparse matrix B
-    double *d_B;
+    float *d_B;
     gpuErrchk(cudaMalloc(&d_B, nnzB * sizeof(*d_B)));
     int *d_B_RowIndices;
     gpuErrchk(cudaMalloc(&d_B_RowIndices, (k + 1) * sizeof(*d_B_RowIndices)));
@@ -85,7 +85,7 @@ void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
     cusparseSafeCall(cusparseDdense2csr(handle, k, n, descrB, d_B_dense, k, d_nnzPerVectorB, d_B, d_B_RowIndices, d_B_ColIndices));
 
     // Move sparse B from device to host
-    // double *h_B = (double *)malloc(nnzB * sizeof(*h_B));
+    // float *h_B = (float *)malloc(nnzB * sizeof(*h_B));
     // int *h_B_ColIndices = (int *)malloc(nnzB * sizeof(*h_B_ColIndices));
     // int *h_B_RowIndices = (int *)malloc((k + 1) * sizeof(*h_B_RowIndices));
 
@@ -117,12 +117,12 @@ void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
     }
 
     // device side sparse matrix C
-    double *d_C;
-    gpuErrchk(cudaMalloc(&d_C, nnzC * sizeof(double)));
+    float *d_C;
+    gpuErrchk(cudaMalloc(&d_C, nnzC * sizeof(float)));
     int *d_C_ColIndices;
     gpuErrchk(cudaMalloc(&d_C_ColIndices, nnzC * sizeof(int)));
     // host side sparse matrix c
-    double *h_C = (double *)malloc(nnzC * sizeof(*h_C));
+    float *h_C = (float *)malloc(nnzC * sizeof(*h_C));
     int *h_C_ColIndices = (int *)malloc(nnzC * sizeof(*h_C_ColIndices));
     int *h_C_RowIndices = (int *)malloc((m + 1) * sizeof(*h_C_RowIndices));
 
@@ -136,7 +136,7 @@ void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
     gpuErrchk(cudaMemcpy(h_C_RowIndices, d_C_RowIndices, (m + 1) * sizeof(*h_C_RowIndices), cudaMemcpyDeviceToHost));
     gpuErrchk(cudaMemcpy(h_C_ColIndices, d_C_ColIndices, nnzC * sizeof(*h_C_ColIndices), cudaMemcpyDeviceToHost));
 
-    gpuErrchk(cudaMemcpy(h_C_dense, d_C_dense, m * n * sizeof(double), cudaMemcpyDeviceToHost));
+    gpuErrchk(cudaMemcpy(h_C_dense, d_C_dense, m * n * sizeof(float), cudaMemcpyDeviceToHost));
 
     for (int j = 0; j < m; j++)
     {
@@ -171,14 +171,14 @@ void cusparse_mm_wrapper(double *h_A, int *h_A_ColIndices, int *h_A_RowIndices,
     return;
 }
 
-void dense_to_csr(double *h_A_dense, const int Nrows, const int Ncols, double **h_A_val, int **h_A_colind, int **h_A_rowptr, int *nnzA)
+void dense_to_csr(float *h_A_dense, const int Nrows, const int Ncols, float **h_A_val, int **h_A_colind, int **h_A_rowptr, int *nnzA)
 {
     // Initialize cuSPARSE
     cusparseHandle_t handle;
     cusparseSafeCall(cusparseCreate(&handle));
 
     //create device array and copy host to it
-    double *d_A_dense;
+    float *d_A_dense;
     gpuErrchk(cudaMalloc(&d_A_dense, Nrows * Ncols * sizeof(*d_A_dense)));
     gpuErrchk(cudaMemcpy(d_A_dense, h_A_dense, Nrows * Ncols * sizeof(*d_A_dense), cudaMemcpyHostToDevice));
 
@@ -199,7 +199,7 @@ void dense_to_csr(double *h_A_dense, const int Nrows, const int Ncols, double **
     gpuErrchk(cudaMemcpy(h_nnzPerVector, d_nnzPerVector, Nrows * sizeof(*h_nnzPerVector), cudaMemcpyDeviceToHost));
 
     // Device side dense matrix
-    double *d_A;
+    float *d_A;
     gpuErrchk(cudaMalloc(&d_A, nnz * sizeof(*d_A)));
     int *d_A_RowIndices;
     gpuErrchk(cudaMalloc(&d_A_RowIndices, (Nrows + 1) * sizeof(*d_A_RowIndices)));
@@ -209,7 +209,7 @@ void dense_to_csr(double *h_A_dense, const int Nrows, const int Ncols, double **
     cusparseSafeCall(cusparseDdense2csr(handle, Nrows, Ncols, descrA, d_A_dense, lda, d_nnzPerVector, d_A, d_A_RowIndices, d_A_ColIndices));
 
     // Host side dense matrix
-    double *h_A = (double *)malloc(nnz * sizeof(*h_A));
+    float *h_A = (float *)malloc(nnz * sizeof(*h_A));
     int *h_A_RowIndices = (int *)malloc((Nrows + 1) * sizeof(*h_A_RowIndices));
     int *h_A_ColIndices = (int *)malloc(nnz * sizeof(*h_A_ColIndices));
     gpuErrchk(cudaMemcpy(h_A, d_A, nnz * sizeof(*h_A), cudaMemcpyDeviceToHost));
@@ -237,7 +237,7 @@ int main()
     // (3 x 4) x (4 x 2) = 3 x 2
     int h_A_rows = 3;
     int h_A_cols = 4;
-    double *h_A_dense = (double *)malloc(h_A_rows * h_A_cols * sizeof(*h_A_dense));
+    float *h_A_dense = (float *)malloc(h_A_rows * h_A_cols * sizeof(*h_A_dense));
 
     // Column-major ordering
     h_A_dense[0] = 1.0f;
@@ -253,7 +253,7 @@ int main()
     h_A_dense[10] = 2.0f;
     h_A_dense[11] = 9.0f;
 
-    double *h_A_val = nullptr;
+    float *h_A_val = nullptr;
     int *h_A_colind = nullptr;
     int *h_A_rowptr = nullptr;
     int nnzA = 0;
@@ -272,7 +272,7 @@ int main()
 
     int h_B_rows = 4;
     int h_B_cols = 2;
-    double *h_B_dense = (double *)malloc(h_B_rows * h_B_cols * sizeof(*h_B_dense));
+    float *h_B_dense = (float *)malloc(h_B_rows * h_B_cols * sizeof(*h_B_dense));
 
     // Column-major ordering
     h_B_dense[0] = 1.0f;
